+++ lib/src/point.h
@@ -4,9 +4,9 @@
 #include "tree_sitter/api.h"
 
 #define POINT_ZERO ((TSPoint) {0, 0})
-#define POINT_MAX ((TSPoint) {UINT32_MAX, UINT32_MAX})
+#define POINT_MAX ((TSPoint) {UINT64_MAX, UINT64_MAX})
 
-static inline TSPoint point__new(unsigned row, unsigned column) {
+static inline TSPoint point__new(uint64_t row, uint64_t column) {
   TSPoint result = {row, column};
   return result;
 }
