+++ lib/src/reduce_action.h
@@ -9,7 +9,7 @@ extern "C" {
 #include "tree_sitter/api.h"
 
 typedef struct {
-  uint32_t count;
+  uint64_t count;
   TSSymbol symbol;
   int dynamic_precedence;
   unsigned short production_id;
@@ -19,7 +19,7 @@ typedef Array(ReduceAction) ReduceActionSet;
 
 static inline void ts_reduce_action_set_add(ReduceActionSet *self,
                                             ReduceAction new_action) {
-  for (uint32_t i = 0; i < self->size; i++) {
+  for (uint64_t i = 0; i < self->size; i++) {
     ReduceAction action = self->contents[i];
     if (action.symbol == new_action.symbol && action.count == new_action.count)
       return;
