+++ lib/src/stack.h
@@ -35,7 +35,7 @@ Stack *ts_stack_new(SubtreePool *);
 void ts_stack_delete(Stack *);
 
 // Get the stack's current number of versions.
-uint32_t ts_stack_version_count(const Stack *);
+uint64_t ts_stack_version_count(const Stack *);
 
 // Get the state at the top of the given version of the stack. If the stack is
 // empty, this returns the initial state, 0.
@@ -62,7 +62,7 @@ void ts_stack_push(Stack *, StackVersion, Subtree , bool, TSStateId);
 // versions which had previously been merged. It returns an array that
 // specifies the index of each revealed version and the trees that were
 // removed from that version.
-StackSliceArray ts_stack_pop_count(Stack *, StackVersion, uint32_t count);
+StackSliceArray ts_stack_pop_count(Stack *, StackVersion, uint64_t count);
 
 // Remove an error at the top of the given version of the stack.
 SubtreeArray ts_stack_pop_error(Stack *, StackVersion);
@@ -124,7 +124,7 @@ void ts_stack_clear(Stack *);
 
 bool ts_stack_print_dot_graph(Stack *, const TSLanguage *, FILE *);
 
-typedef void (*StackIterateCallback)(void *, TSStateId, uint32_t);
+typedef void (*StackIterateCallback)(void *, TSStateId, uint64_t);
 
 #ifdef __cplusplus
 }
